+++ src/components/myPage/Statistics/StatsByProvider.jsx
@@ -1,7 +1,7 @@
 import { Progress, Box, Text, Group, SimpleGrid, useMantineColorScheme } from '@mantine/core';
 import { useMediaQuery } from '@mantine/hooks';
 import styled from '@emotion/styled';
-import { useStatsByProvider } from '../../hooks/statistics';
+import { useStatsByProvider } from '../../../hooks/statistics';
 
 const ProgressLabel = styled(Progress)`
   .mantine-Progress-label {
