+++ src/components/myPage/Statistics/index.js
@@ -1,0 +1,6 @@
+export { default as Statistics } from './Statistics';
+export { default as StatsWrapper } from './StatsWrapper';
+export { default as StatsByProvider } from './StatsByProvider';
+export { default as StatsByMonthly } from './StatsByMonthly';
+// export { default as StatsByGenre } from './StatsByGenre';
+export { default as StatsByGenre } from '../StatsByGenre';
